Improve error messages
This is definitely a nitpick and mostly my own opinion.

"option requires a value: ..." makes it sound like the option requires
itself

"option does not require a value: ..." makes it sound like the value is
optional

Since the text will always be something the programmer has agreed is an
acceptable option. It should be fine to just embed it in the middle of
the message.