Optionally ignore negative numbers as options
This emulates the behaviour of python's argparse when you specify
options that look like negative numbers in your parser (e.g. -1 or -.5
or -123.456 which are not all short options but argparse is versatile).

This is useful if you take a optionally negative number as a positional
argument.

The default is to keep the normal behaviour with the alternative
optionally enabled by calling
Options::allow_number_short_options(..., false)