Forced UTF-8 options
Instead of letting you choose what type you want to get back from the
API, just make it a `&str`. This means that options can only ever be
something which can be turned into valid UTF-8. This sounds restrictive
but for users of `&str as Argument` it makes no difference and for the
few users of `&[u8] as Argument` it simplifies a lot of things.

It also makes `&[u8]` usable on windows via `OsStr::as_encoded_bytes`
and makes `&OsStr as Argument` a possibility.